create cloudprovider "providers" package     move all providers into new package     update all references to old package path
