+++ cmd/kubelet/app/plugins.go
@@ -37,13 +37,7 @@ import (
 	"k8s.io/kubernetes/pkg/volume/rbd"
 	"k8s.io/kubernetes/pkg/volume/secret"
 	//Cloud providers
-	_ "k8s.io/kubernetes/pkg/cloudprovider/aws"
-	_ "k8s.io/kubernetes/pkg/cloudprovider/gce"
-	_ "k8s.io/kubernetes/pkg/cloudprovider/mesos"
-	_ "k8s.io/kubernetes/pkg/cloudprovider/openstack"
-	_ "k8s.io/kubernetes/pkg/cloudprovider/ovirt"
-	_ "k8s.io/kubernetes/pkg/cloudprovider/rackspace"
-	_ "k8s.io/kubernetes/pkg/cloudprovider/vagrant"
+	_ "k8s.io/kubernetes/pkg/cloudprovider/providers"
 )
 
 // ProbeVolumePlugins collects all volume plugins into an easy to use list.
