--- pkg/cloudprovider/openstack/MAINTAINERS.md
@@ -1,6 +1,0 @@
-# Maintainers
-
-* [Angus Lees](https://github.com/anguslees)
-
-
-[![Analytics](https://kubernetes-site.appspot.com/UA-36037335-10/GitHub/pkg/cloudprovider/openstack/MAINTAINERS.md?pixel)]()
