+++ pkg/cloudprovider/providers/gce/token_source.go
