+++ pkg/cloudprovider/providers/rackspace/MAINTAINERS.md
@@ -1,0 +1,6 @@
+# Maintainers
+
+* [Thom May](https://github.com/thommay)
+
+
+[![Analytics](https://kubernetes-site.appspot.com/UA-36037335-10/GitHub/pkg/cloudprovider/providers/rackspace/MAINTAINERS.md?pixel)]()
