+++ pkg/cloudprovider/providers/rackspace/rackspace.go
