+++ pkg/cloudprovider/providers/rackspace/rackspace_test.go
