+++ pkg/cloudprovider/providers/vagrant/doc.go
