+++ pkg/cloudprovider/providers/vagrant/vagrant.go
