+++ pkg/cloudprovider/providers/vagrant/vagrant_test.go
