--- pkg/cloudprovider/rackspace/MAINTAINERS.md
@@ -1,6 +1,0 @@
-# Maintainers
-
-* [Thom May](https://github.com/thommay)
-
-
-[![Analytics](https://kubernetes-site.appspot.com/UA-36037335-10/GitHub/pkg/cloudprovider/rackspace/MAINTAINERS.md?pixel)]()
