+++ pkg/controller/service/servicecontroller_test.go
@@ -22,7 +22,7 @@ import (
 
 	"k8s.io/kubernetes/pkg/api"
 	"k8s.io/kubernetes/pkg/client/testclient"
-	fake_cloud "k8s.io/kubernetes/pkg/cloudprovider/fake"
+	fake_cloud "k8s.io/kubernetes/pkg/cloudprovider/providers/fake"
 	"k8s.io/kubernetes/pkg/types"
 )
